fix creazione e cancellazione progetto e definzioni api correlate

+++ src/services/project.service.ts
@@ -1,12 +1,12 @@
-import { Repository } from "typeorm";
+import { Evaluation } from "./../entity/evaluation.entity";
+import { FindOptionsWhere, Repository } from "typeorm";
 import { AppDataSource } from "../database/database";
 import { Project } from "../entity/project.entity";
 import { ProjectResponse } from "../models/project.model"; // Supponiamo che ProjectResponse sia un DTO
 import { User } from "../entity/user.entity";
-import { EvaluationService } from "./evalutaion.service";
+import { EvaluationService } from "./evaluation.service";
 import { Value } from "../entity/values.entity";
 import { Skill } from "../entity/skill.entity";
-import { Evaluation } from "../entity/evaluation.entity";
 import e from "express";
 
 export class ProjectService {
@@ -30,53 +30,58 @@ export class ProjectService {
     try {
       const user = await this.userRepository.findOne({
         where: { id: userId },
-        relations: ["projects"], // Assicurati che la relazione "projects" sia caricata
+        relations: [
+          "projects",
+          "projects.skills",
+          "projects.evaluation",
+          "projects.evaluation.values",
+          "projects.evaluation.values.skill",
+        ],
       });
-      // Verifica se l'utente esiste
+
       if (!user || !user.projects) {
         throw new Error("L'utente non esiste o non ha progetti associati.");
       }
 
-      // Ottieni i progetti associati all'utente
       const projects = user.projects;
 
-      // Mappa i progetti e ottieni la valutazione più recente per ciascuno
       const projectResponses = await Promise.all(
         projects.map(async (project: Project) => {
-          // Ottieni la valutazione più recente per il progetto e l'utente
-          const latestEvaluation =
-            await this.evaluationService.getLatestEvaluation(
-              project.id,
-              userId
-            );
-
+          console.log(project);
           let response: ProjectResponse = {
-            id: project.id.toString(),
+            id: project.id ? project.id.toString() : "",
             projectName: project.name,
             ratingAverage: 0,
+            labelEvaluations: project.skills
+              ? project.skills.map((skill: Skill) => {
+                  let label = {
+                    id: skill.id.toString(),
+                    label: skill.name,
+                    shortLabel: skill.shortName,
+                  };
+                  return label;
+                })
+              : [],
+            evaluations: project.evaluation
+              ? project.evaluation.map((evaluation: Evaluation) => {
+                  let eva = {
+                    id: evaluation.id.toString(),
+                    label: evaluation.evaluationDate.toLocaleDateString(), //01/01/2024
+                    values: evaluation.values
+                      ? evaluation.values.map((value: Value) => {
+                          let v = {
+                            id: value.id?.toString(),
+                            skill: value.skill.name,
+                            value: value.value,
+                          };
+                          return v;
+                        })
+                      : [],
+                  };
+                  return eva;
+                })
+              : [],
           };
-          // Se non ci sono valutazioni, ritorna `null`
-          if (!latestEvaluation) return response;
-          console.log(latestEvaluation);
-          response.evaluations = [
-            {
-              id: latestEvaluation.id.toString(),
-              label: latestEvaluation.evaluationDate.toString(), // La data potrebbe essere rappresentata come stringa
-              values:
-                latestEvaluation.values?.map((value: Value) => ({
-                  id: value.id.toString(),
-                  skill: value.skill.name,
-                  value: value.value,
-                })) ?? [],
-            },
-          ];
-          response.labelEvaluations =
-            latestEvaluation.values?.map((label: Value) => ({
-              id: label.id.toString(),
-              label: label.skill.name,
-              shortLabel: label.skill.shortName,
-            })) ?? [];
-
           return response;
         })
       );
@@ -90,23 +95,11 @@ export class ProjectService {
     }
   }
 
-  // Metodo per calcolare la valutazione media
-  private calculateRatingAverage(values: Value[]): number {
-    // Calcola la somma totale delle valutazioni
-    const totalRating = values.reduce((acc, value) => acc + value.value, 0);
-
-    // Il conteggio delle valutazioni
-    const totalCount = values.length;
-
-    // Restituisce la media delle valutazioni o 0 se non ci sono valori
-    return totalCount > 0 ? totalRating / totalCount : 0;
-  }
-
-  // **Read (One)**: Ottieni un progetto per ID
   async getProjectById(id: number): Promise<Project | null> {
     try {
       const project = await this.projectRepository.findOne({
         where: { id },
+        relations: ["users", "skills", "evaluations"],
       });
       if (!project) {
         throw new Error("Progetto non trovato.");
@@ -118,18 +111,16 @@ export class ProjectService {
     }
   }
 
-  // **Update**: Aggiorna un progetto esistente
   async updateProject(
     id: number,
     updatedProject: Partial<Project>
   ): Promise<Project> {
     try {
-      const project = await this.getProjectById(id); // Ottieni il progetto
+      const project = await this.getProjectById(id);
       if (!project) {
         throw new Error("Progetto non trovato.");
       }
-      // Aggiorna il progetto con i dati nuovi
-      Object.assign(project, updatedProject); // Copia i nuovi valori nel progetto esistente
+      Object.assign(project, updatedProject);
       return await this.projectRepository.save(project);
     } catch (error) {
       console.error("Errore durante l'aggiornamento del progetto:", error);
@@ -137,15 +128,19 @@ export class ProjectService {
     }
   }
 
-  async createProject(project: Partial<Project>): Promise<Project> {
-    let newProject: Project;
+  async createProject(project: Partial<Project>): Promise<number> {
     try {
+      let newProject: Project = this.projectRepository.create();
       const user = await this.userRepository.findOne({
         where: { id: 1 },
-        relations: ["projects"], // Assicurati che la relazione "projects" sia caricata
+        relations: ["projects"],
       });
+      if (!user) {
+        throw new Error("Utente non trovato!");
+      }
+      project.name = "Test";
+      project.users = [user];
 
-      project.name = "Progetto di prova";
       let skill1 = await this.skillRepository.findOne({
         where: { name: "JavaScript" },
       });
@@ -154,7 +149,7 @@ export class ProjectService {
           name: "JavaScript",
           shortName: "JS",
         });
-        await this.skillRepository.save(skill1);
+        skill1 = await this.skillRepository.save(skill1);
       }
 
       let skill2 = await this.skillRepository.findOne({
@@ -165,48 +160,47 @@ export class ProjectService {
           name: "TypeScript",
           shortName: "TS",
         });
-        await this.skillRepository.save(skill2);
+        skill2 = await this.skillRepository.save(skill2);
       }
 
       project.skills = [skill1, skill2];
-
       newProject = await this.projectRepository.save(project);
-      if (!user) {
-        throw new Error("Utente non trovato!");
-      }
-      let value: Value = this.valueRepository.create({
-        skill: skill1,
-        value: 10,
-      } as Partial<Value>);
-      await this.valueRepository.save(value);
 
-      let evaluation: Evaluation = this.evaluationRepository.create({
-        project: newProject,
-        user: user,
-        values: [value],
+      let evaluation = this.evaluationRepository.create({
         evaluationDate: new Date(),
+        user: user,
+        project: newProject,
       });
       evaluation = await this.evaluationRepository.save(evaluation);
 
-      // Associa l'utente al progetto
-      user.projects = [...(user.projects || []), newProject];
-      await this.userRepository.save(user); // ✅ Salva l'utente con il progetto associato
+      let value = this.valueRepository.create({
+        skill: skill1,
+        value: 10,
+        evaluation: evaluation,
+      });
+      await this.valueRepository.save(value);
+      let value1 = this.valueRepository.create({
+        skill: skill2,
+        value: 5,
+        evaluation: evaluation,
+      });
+      await this.valueRepository.save(value1);
+
+      return newProject.id;
     } catch (error) {
       console.error("Errore durante l'aggiornamento del progetto:", error);
       throw new Error("Non è stato possibile aggiornare il progetto.");
     }
-
-    return newProject;
   }
 
   // **Delete**: Elimina un progetto per ID
   async deleteProject(id: number): Promise<void> {
     try {
-      const project = await this.getProjectById(id);
-      if (!project) {
-        throw new Error("Progetto non trovato.");
-      }
-      await this.projectRepository.remove(project); // Elimina il progetto
+      const project = await this.projectRepository.findOne({
+        where: { id },
+      });
+      if (project) await this.projectRepository.remove(project);
+      console.log(`Progetto con id ${id} eliminato con successo!`);
     } catch (error) {
       console.error("Errore durante l'eliminazione del progetto:", error);
       throw new Error("Non è stato possibile eliminare il progetto.");
